refactor: move RecordBatch interfaces and classes to api package
