+++ engine/src/main/java/io/camunda/zeebe/engine/api/ProcessingResult.java
@@ -7,8 +7,9 @@
  */
 package io.camunda.zeebe.engine.api;
 
+import io.camunda.zeebe.engine.api.records.ImmutableRecordBatch;
+import io.camunda.zeebe.engine.api.records.RecordBatchEntry;
 import io.camunda.zeebe.logstreams.log.LogStreamBatchWriter;
-import io.camunda.zeebe.streamprocessor.records.ImmutableRecordBatch;
 
 /**
  * Here the interface is just a suggestion. Can be whatever PDT teams thinks is best to work with
@@ -20,7 +21,7 @@ public interface ProcessingResult {
 
   /**
    * Returns the resulting record batch, which can be empty or consist of multiple {@link
-   * io.camunda.zeebe.streamprocessor.records.RecordBatchEntry}s. These entries are the result of
+   * RecordBatchEntry}s. These entries are the result of
    * the current processing. If an entry is of type {@link
    * io.camunda.zeebe.protocol.record.RecordType#COMMAND} it will be later processed as follow-up
    * command.
