+++ engine/src/main/java/io/camunda/zeebe/engine/api/records/RecordBatchSizePredicate.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.streamprocessor.records;
+package io.camunda.zeebe.engine.api.records;
 
 import java.util.function.BiPredicate;
