+++ engine/src/main/java/io/camunda/zeebe/streamprocessor/DirectProcessingResult.java
@@ -11,8 +11,8 @@ import io.camunda.zeebe.engine.api.CommandResponseWriter;
 import io.camunda.zeebe.engine.api.PostCommitTask;
 import io.camunda.zeebe.engine.api.ProcessingResult;
 import io.camunda.zeebe.engine.api.TaskResult;
+import io.camunda.zeebe.engine.api.records.ImmutableRecordBatch;
 import io.camunda.zeebe.logstreams.log.LogStreamBatchWriter;
-import io.camunda.zeebe.streamprocessor.records.ImmutableRecordBatch;
 import java.util.ArrayList;
 import java.util.List;
