+++ engine/src/main/java/io/camunda/zeebe/streamprocessor/DirectTaskResultBuilder.java
@@ -9,11 +9,11 @@ package io.camunda.zeebe.streamprocessor;
 
 import io.camunda.zeebe.engine.api.TaskResult;
 import io.camunda.zeebe.engine.api.TaskResultBuilder;
+import io.camunda.zeebe.engine.api.records.RecordBatch;
 import io.camunda.zeebe.protocol.record.RecordType;
 import io.camunda.zeebe.protocol.record.RecordValue;
 import io.camunda.zeebe.protocol.record.RejectionType;
 import io.camunda.zeebe.protocol.record.intent.Intent;
-import io.camunda.zeebe.streamprocessor.records.RecordBatch;
 import java.util.Collections;
 
 /**
